refactor: 자동차 이름 중복 검증 로직 위치 변경
기존 자동차 목록 객체에서 입력 검증 부분으로 위치를 변경합니다.
